Short introduction to this project on Github.
Provide link to the recent findings. CWEB 2.7 and 2.8 are available in
Utah. Spotted CWEB 2.8 on the Stanford FTP server.

Spotted CWEB 3.0 in an old Atari package.  Requires 'zoo' to extract.
CWEB 3.0 is contained as 'original.lzh', which requires 'lha' to
uncommpress. Weird.

Spotted CWEB 3.1 in an old Amiga package. Also requires 'lha' to
extract.