Correction of comment about 'out_ptr'.
It seems that Igor Liferenko is right in pointing out that 'out_ptr'
denotes the position of the last character in 'out_buf'. To make things
clearer the pointer variable is initialized at the point of declaration
(RAII) and the 'Set init...' section is slightly simplified by using the
'out()' macro with the first character to be put into the buffer.