Fix '\pdfURL' macro.
Contrary to 'point 18' in the CWEB manual, '\pdfURL' could only be used
with some form of 'pdf TeX', not with 'plain TeX'. The macro itself had
provisions for all cases, so it can safely be moved out of the '\ifacro'
wrapper. Also, the two helper macros for 'tilde' and 'underscore' can be
simplified by re-using formerly defined stuff.