Extract common section code.
CTANGLE.W and CWEAVE.W have several sections with (almost) identical
CWEB code. For this experiment, we extract these sections into common.h
to eliminate redundancy.

Note that these sections are not used in COMMON.W, so there are several
warnings at this time. (Maybe a new include file is in order.)