[Issues #58 and #59] Don't break CWEAVE.
Inside 'output_defs' we can't use standard 'isspace', because '++' is
mapped to 'plus_plus 013' (section 6 from common.h), which is '\cr' in
the ASCII table and so 'isspace(013)' is true.

Apart from  macro 'gather_digits_while' in CWEAVE, it would be far too
likely that other macros end in similar form and get clobbered.